requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
django==4.2.6
    # via
    #   -r requirements.in
    #   django-admin-inline-paginator
    #   django-js-asset
django-admin-inline-paginator==0.4.0
    # via -r requirements.in
django-js-asset==2.1.0
    # via django-mptt
django-mptt==0.15.0
    # via -r requirements.in
django-nested-admin==4.0.2
    # via -r requirements.in
python-monkey-business==1.0.0
    # via django-nested-admin
six==1.16.0
    # via python-monkey-business
sqlparse==0.4.4
    # via django
